requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
anytree==2.13.0
APScheduler==3.11.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
backend==0.2.4.1
backrefs==5.8
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssutils==2.11.1
decorator==5.2.1
deepface==0.0.93
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
dotenv==0.9.9
ecdsa==0.19.1
EditorConfig==0.17.0
executing==2.2.0
fastapi==0.115.0
fastjsonschema==2.21.1
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.0
flatbuffers==25.2.10
fsspec==2025.3.2
gast==0.6.0
gdown==5.2.0
generativeai==0.0.1
ghp-import==2.1.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.20.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.0
grpcio==1.71.0
grpcio-status==1.71.0
gspread==6.2.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
iniconfig==2.1.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsbeautifier==1.15.4
json_repair==0.29.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kaggle==1.7.4.5
keras==3.9.2
lark-parser==0.12.0
libclang==18.1.1
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mergedeep==1.3.4
mistune==3.1.3
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.14
mkdocs-material-extensions==1.3.1
ml_dtypes==0.5.1
more-itertools==10.6.0
mpmath==1.3.0
mtcnn==1.0.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
ngrok==1.4.0
numpy==2.1.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
openai==1.84.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
outcome==1.3.0.post0
packaging==24.2
paginate==0.5.7
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.8
playwright==1.51.0
pluggy==1.6.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.1.1
Pygments==2.19.1
pymdown-extensions==10.15
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
selenium==4.31.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.38.6
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
text-unidecode==1.3
tf_keras==2.19.0
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.3
trio==0.29.0
trio-websocket==0.12.2
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.30.6
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
yarg==0.1.9
